[feat] 서버와 통신하여 게시글 생성 구현 (#8)
* [feat] 게시글 작성 api통신 및 db등록 구현

* [feat] 취소버튼 리다이렉션 구현

* [fix] create라우터 수정
