+++ src/App.js
@@ -1,6 +1,7 @@
 import Main from "./pages/Main";
 import Posts from "./pages/Posts";
 import Post from "./pages/Post";
+import CreatePost from "./pages/CreatePost";
 import Game from "./pages/Game";
 import { Routes, Route } from "react-router";
 import "./App.css";
@@ -11,6 +12,7 @@ function App() {
       <Route path="/" element={<Main />} />
       <Route path="/posts" element={<Posts />} />
       <Route path="/posts/:id" element={<Post />} />
+      <Route path="/posts/create" element={<CreatePost />} />
       <Route path="/game" element={<Game />} />
     </Routes>
   );
