+++ src/pages/Posts.js
@@ -1,20 +1,38 @@
-import ReadPosts from "../components/ReadPosts";
-import CreatePost from "../components/CreatePost";
-import { useState } from "react";
+import { Link } from "react-router-dom";
+import Grade from "../components/Grade";
+import PastTime from "../components/PastTime";
+import { useState, useEffect } from "react";
 
 function Posts() {
-  const [mode, setMode] = useState("read");
+  const [posts, setPosts] = useState(null);
 
+  useEffect(() => {
+    fetch("http://localhost:3001/posts")
+      .then((response) => response.json())
+      .then((result) =>
+        setPosts(() =>
+          result.map((post) => (
+            // css생성 후 style속성 삭제 필요
+            <Link
+              to={`/posts/${post._id}`}
+              key={post._id}
+              style={{ display: "flex" }}
+            >
+              <div>{post.title}</div>
+              <Grade grade={post.grade} />
+              <div>{post.author}</div>
+              <PastTime createdAt={post.createdAt} />
+            </Link>
+          ))
+        )
+      );
+  }, []);
   return (
     <div>
-      {mode === "read" && (
-        <div>
-          <ReadPosts />
-          <button onClick={() => setMode("create")}>작성하기</button>
-        </div>
-      )}
-
-      {mode === "create" && <CreatePost />}
+      {posts}
+      <Link to={"create"}>
+        <button>작성하기</button>
+      </Link>
     </div>
   );
 }
